+++ internal/apm/command_test.go
