+++ internal/config/table_renderer.go
