+++ internal/credentials/command_test.go
